+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## [2.0.0](https://github.com/smartiniOnGitHub/fastify-webhook/releases/tag/2.0.0) (unreleased)
+[Full Changelog](https://github.com/smartiniOnGitHub/fastify-webhook/compare/1.0.0...2.0.0)
+Summary Changelog:
+- Update requirements to Fastify v2
+- Update all dependencies
+- Breaking Change: rename plugin option 'beforeHandler' to 'preHandler', 
+  due to a change/rename in Fastify v2
+- Update function arguments like in Fastify v2 (from req to request, etc)
+
 ## [1.0.0](https://github.com/smartiniOnGitHub/fastify-webhook/releases/tag/1.0.0) (2019-04-05)
 Summary Changelog:
 - Updated all dependencies